valueverse/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.12.1
redis==5.0.1
celery==5.3.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
httpx==0.25.2
langchain==0.1.0
openai==1.6.1
anthropic==0.8.1
pinecone-client==2.2.4
websockets==12.0
python-socketio==5.10.0
aiofiles==23.2.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
ruff==0.1.8
